Fusion bug: don't lose arguments to Done instruction
The continuation for the Done instruction doesn't usually take arguments,
because it can't do anything with them. But if it does take arguments for
some weird reason, the fusion algorithm needs to keep them around.
Before, fusion was removing the arguments from the update, but leaving
them in the continuation binding, making the continuation under-applied.