use bounded channels between processes
We really need some mechanism for fairness between channels, so bounded
channels are fine here. It would be nice to try something like Parks or
other KPN methods which introduce a soft upper bound which you can
increase as necessary.

At the moment this is a single element, so maybe a simple MVar would be
better, but I think we'd be better off adding a bound parameter instead.